Adds extension and UDF pages to admin menu
Fixes #N/A

Adds JavaScript file for mapping page and updates select option label

This commit adds a JavaScript file for the mapping page in the admin section, which enables AJAX functionality for field mapping. It also updates the label of a select option to provide a more accurate description for the default option.

Fixes #N/A