Refactors SAP synchronization handling in class-sap-sync.php
Improves SAP synchronization logic in the SAPWC_Sync_Handler class by refactoring the code. Updates the retrieval of the card code and CIF from user metadata, ensuring proper handling of array values. Simplifies the building of the payload by extracting and trimming relevant information from the order object. Modifies the values of $u_ruta and $u_portes. Adds optional fields to the payload if corresponding values are not empty. Enhances the clarity and maintainability of the code.

Fixes #123

Refactors SAP synchronization handling in class-sap-sync.php

Improves the logic of SAP synchronization in the SAPWC_Sync_Handler class. Refactors the code to enhance clarity and maintainability. Updates the retrieval of the card code and CIF from user metadata. Simplifies payload building by extracting and trimming relevant information from the order object. Modifies the values of $u_ruta and $u_portes. Adds optional fields to the payload if corresponding values are not empty.

Fixes #123