Refactors user and salesperson retrieval logic
Improves the logic for retrieving users and salespersons in the `class-sync-options-page.php` file. Filters out locked users and adds a check to mark users who are also salespersons. Sorts the users alphabetically. Updates the description for the user sign input field.

Updates the stable tag in the `readme.txt` file to version 1.2.31.

Fixes #789